pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "slidesmd"
version = "0.3.1"
description = "Auto-index PowerPoint presentations into a central agents.md — queryable by any AI"
authors = ["Chanukya Pekala <chanukya.pekala@gmail.com>"]
readme = "README.md"
packages = [{include = "slidesmd"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
watchdog = ">=4.0.0"
python-pptx = ">=1.0.0"
typer = ">=0.12.0"
rich = ">=13.0.0"
pytesseract = "^0.3.13"
fastmcp = ">=2.0.0"
sentence-transformers = {version = ">=2.2.0", optional = true}
numpy = {version = ">=1.24.0", optional = true}

[tool.poetry.extras]
semantic = ["sentence-transformers", "numpy"]

[tool.poetry.scripts]
slidesmd = "slidesmd.cli:app"
slidesmd-mcp = "slidesmd.mcp_server:main"

[tool.poetry.group.dev.dependencies]
ruff = "^0.15.11"
pytest = ">=7.0"

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true